Traductor/requirements (1).txt
streamlit==1.31.0
gtts==2.5.0
googletrans==4.0.0rc1
PyPDF2==3.0.1
python-docx==1.1.0
